Include examples in released source packages.
Older versions of cargo read all of Cargo.toml and search for the paths
of examples mentioned in it, even if the Cargo.toml is for a library in
the registry. This was changed in some recent release, but I'm not sure
when.

TODO: add an old-rustc test to the continuous integration to avoid
situations such as this.