Rework DateBasedLogFile a bit in my own style.
I sincerely apologize to the original author for my not handling this
better. I was swamped when I got the PR, and didn't properly review it
nor give good feedback. I gave a bit, but didn't go in-depth enough to
come up with the best quality implementation we could together.

I respect your original contribution here, but I don't want to bother
you further for more and more review over 4 months since you originally
opened your Pull Request. Unfortunately, I also don't want to commit
to a sub-par API and implementation in fern.

The original work was good, but it was not the best it could be. I've
changed some of the design decisions for flexibility, and some of the
implementation for what I believe to be better clarity and efficiency.

And of course, not all of these changes are because of "sub-par"
code. Most are just stylistic, and some are API design choices which
I believe are more intuitive (like not inserting a separator between
prefix and suffix, so that more schemes can be supported). Other
changes are arbitrary decisions that I think could cause slightly less
confusion, like checking the log file before logging rather than after.

Hopefully this isn't too bad of an offense on my part.